refactor(snapshot): simplify snapshotctl watch wait tests
